+++ flogin/default_events.py
@@ -26,7 +26,7 @@ async def on_error(
 
 def get_default_events(plugin: Plugin[Any]) -> dict[str, Callable[..., Awaitable[Any]]]:
     def on_query(data: dict[str, Any], raw_settings: dict[str, Any]):
-        query = Query.from_json(data)
+        query = Query(data, plugin)
         if plugin._settings_are_populated is False:
             LOG.info(f"Settings have not been populated yet, creating a new instance")
             plugin._settings_are_populated = True