+++ docs/whats_new.rst
@@ -14,6 +14,8 @@ Breaking Changes
     - Remove ``flogin.errors.SettingNotFound``
 - Rewrite the CLI commands
 - Remove ``flogin.conditions.MultiCondition`` in favor of :class:`flogin.conditions.AnyCondition` and :class:`flogin.conditions.AllCondition`
+- Remove `Query.from_json`
+- For :func:`flogin.testing.plugin_tester.PluginTester.test_query`, switch from receiving a query object to taking kwargs that will be used to make a query object
 - Rename the ``flogin.flow_api`` directory to ``flogin.flow``
     - Rename ``flogin.flow_api.client.py`` to ``flogin.flow.api.py``
 
@@ -33,6 +35,8 @@ New Features
 - Make :attr:`flogin.jsonrpc.results.Result.title` optional
 - Add :class:`flogin.conditions.AnyCondition`
 - Add :class:`flogin.conditions.AllCondition`
+- Add :func:`flogin.query.Query.update_results`
+- Add :func:`flogin.query.Query.update`
 - Add ``flogin.flow.settings.py``
     - Add :class:`flogin.flow.settings.CustomFileManager`
     - Add :class:`flogin.flow.settings.CustomBrowser`